requirements.txt.txt
Flask==2.3.3
psycopg2-binary==2.9.6
Werkzeug==2.3.7
SQLAlchemy==2.0.20
Flask-WTF==1.1.1
python-dotenv==1.0.0